event handlers [nfc]: Factor out connectivity change handler.
Pull the NetInfo event transformer out into its own mini-component.

This is notable for being the first use of React Hooks (introduced in
React Native 0.59) in this project. It may also be the only function-
component.
